Complete completeNotification endpoint
