+++ kifi-backend/modules/common/app/com/keepit/eliza/ElizaServiceClient.scala
@@ -10,7 +10,7 @@ import com.keepit.common.routes.Eliza
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.net.{ CallTimeouts, HttpClient }
 import com.keepit.common.zookeeper.ServiceCluster
-import com.keepit.notify.model.GroupingNotificationKind
+import com.keepit.notify.model.{Recipient, GroupingNotificationKind}
 import com.keepit.notify.model.event.NotificationEvent
 import com.keepit.search.index.message.ThreadContent
 import com.keepit.common.cache.TransactionalCaching.Implicits.directCacheAccess
@@ -75,7 +75,7 @@ trait ElizaServiceClient extends ServiceClient {
 
   def sendNotificationEvent(event: NotificationEvent): Future[Unit]
 
-  def completeNotification[N <: NotificationEvent, G](kind: GroupingNotificationKind[N, G], params: G): Future[Boolean]
+  def completeNotification[N <: NotificationEvent, G](kind: GroupingNotificationKind[N, G], params: G, recipient: Recipient): Future[Boolean]
 
   def getThreadContentForIndexing(sequenceNumber: SequenceNumber[ThreadContent], maxBatchSize: Long): Future[Seq[ThreadContent]]
 
@@ -168,7 +168,17 @@ class ElizaServiceClientImpl @Inject() (
     call(Eliza.internal.sendNotificationEvent(), payload).imap(_ => ())
   }
 
-  def completeNotification[N <: NotificationEvent, G](kind: GroupingNotificationKind[N, G], params: G): Future[Boolean] = ???
+  def completeNotification[N <: NotificationEvent, G](kind: GroupingNotificationKind[N, G], params: G, recipient: Recipient): Future[Boolean] = {
+    val groupIdentifier = kind.gid.serialize(params)
+    val payload = Json.obj(
+      "recipient" -> recipient,
+      "kind" -> kind,
+      "groupIdentifier" -> groupIdentifier
+    )
+    call(Eliza.internal.completeNotification(), payload).map { resp =>
+       Json.parse(resp.body).as[Boolean]
+    }
+  }
 
   def getThreadContentForIndexing(sequenceNumber: SequenceNumber[ThreadContent], maxBatchSize: Long): Future[Seq[ThreadContent]] = {
     call(Eliza.internal.getThreadContentForIndexing(sequenceNumber, maxBatchSize), callTimeouts = longTimeout)
