--- kifi-backend/modules/eliza/app/com/keepit/notify/NotificationProcessing.scala
@@ -1,110 +1,0 @@
-package com.keepit.notify
-
-import com.google.inject.Inject
-import com.keepit.common.db.Id
-import com.keepit.common.db.slick.Database
-import com.keepit.common.logging.Logging
-import com.keepit.eliza.model._
-import com.keepit.notify.delivery.WsNotificationDelivery
-import com.keepit.notify.info.StandardNotificationInfo
-import com.keepit.notify.model.{GroupingNotificationKind, NotificationKind}
-import com.keepit.notify.model.event.NotificationEvent
-
-import scala.concurrent.ExecutionContext
-
-class NotificationProcessing @Inject() (
-    db: Database,
-    notificationRepo: NotificationRepo,
-    notificationItemRepo: NotificationItemRepo,
-    delivery: WsNotificationDelivery,
-    legacyNotificationCheck: LegacyNotificationCheck,
-    implicit val executionContext: ExecutionContext) extends Logging {
-
-  private def shouldGroupWith(event: NotificationEvent, items: Set[NotificationItem]): Boolean = {
-    val res = event.kind.shouldGroupWith(event, items.map(_.event).asInstanceOf[Set[event.N]])
-    log.info(s"notif_debug for event $event with items $items, decided to group? $res")
-    res
-  }
-
-  private def getGroupIdentifier(event: NotificationEvent): Option[String] = {
-    event.kind match {
-      case kind: GroupingNotificationKind[event.N, _] => Some(kind.gid.serialize(kind.getIdentifier(event)))
-      case _ => None
-    }
-  }
-
-  private def saveToExistingNotification(notifId: Id[Notification], event: NotificationEvent): Notification = {
-    db.readWrite { implicit session =>
-      notificationItemRepo.save(NotificationItem(
-        notificationId = notifId,
-        kind = event.kind,
-        event = event,
-        eventTime = event.time
-      ))
-      val notif = notificationRepo.get(notifId)
-      notificationRepo.save(notif.copy(lastEvent = event.time))
-      notificationRepo.get(notifId)
-    }
-  }
-
-  private def createNewNotification(event: NotificationEvent): Notification = {
-    db.readWrite { implicit session =>
-      val notif = notificationRepo.save(Notification(
-        recipient = event.recipient,
-        kind = event.kind,
-        groupIdentifier = None,
-        lastEvent = event.time
-      ))
-      notificationItemRepo.save(NotificationItem(
-        notificationId = notif.id.get,
-        kind = event.kind,
-        event = event,
-        eventTime = event.time
-      ))
-      notificationRepo.get(notif.id.get)
-    }
-  }
-
-  def processNewEvent(event: NotificationEvent, tryDeliver: Boolean = true): Option[NotificationWithItems] = {
-    val groupIdentifier = getGroupIdentifier(event)
-    val recipient = event.recipient
-    if (legacyNotificationCheck.checkUserExperiment(recipient).experimentEnabled) {
-      val notif = groupIdentifier match {
-        case Some(identifier) =>
-          db.readOnlyMaster { implicit session =>
-            notificationRepo.getByGroupIdentifier(event.recipient, event.kind, identifier)
-          } match {
-            case Some(notifGrouped) => saveToExistingNotification(notifGrouped.id.get, event)
-            case None => createNewNotification(event)
-          }
-        case None =>
-          db.readOnlyMaster { implicit session =>
-            notificationRepo.getLastByRecipientAndKind(event.recipient, event.kind)
-          } match {
-            case Some(existingNotif) =>
-              val notifItems = db.readOnlyMaster { implicit session =>
-                notificationItemRepo.getAllForNotification(existingNotif.id.get)
-              }
-              if (shouldGroupWith(event, notifItems.toSet)) {
-                saveToExistingNotification(existingNotif.id.get, event)
-              } else {
-                createNewNotification(event)
-              }
-            case _ => createNewNotification(event)
-          }
-      }
-      val items = db.readOnlyMaster { implicit session =>
-        notificationItemRepo.getAllForNotification(notif.id.get)
-      }.toSet
-      if (tryDeliver) {
-        legacyNotificationCheck.ifUserExperiment(notif.recipient) { recipient =>
-          delivery.deliver(recipient, NotificationWithItems(notif, items))
-        }
-      }
-      Some(NotificationWithItems(notif, items))
-    } else {
-      None
-    }
-  }
-
-}
