+++ kifi-backend/modules/eliza/conf/elizaService.routes
@@ -84,6 +84,7 @@ GET  /internal/eliza/getUnreadNotifications     @com.keepit.eliza.controllers.in
 POST /internal/eliza/disableDevice              @com.keepit.eliza.controllers.internal.ElizaController.disableDevice(id: Id[com.keepit.realtime.Device])
 
 POST /internal/eliza/notifications/post              @com.keepit.eliza.controllers.internal.NotificationController.postEvent
+POST /internal/eliza/notifications/complete           @com.keepit.eliza.controllers.internal.NotificationController.completeNotification
 
 ##########################################
 # IMPORTS
